refactor: rename ProviderWrapper to AccountProviderWrapper and update related imports
